requirements.txt
google-adk>=1.0.0
google-cloud-aiplatform>=1.60.0
google-cloud-storage>=2.18.0
google-cloud-logging>=3.10.0
opentelemetry-exporter-gcp-trace>=1.6.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pandas>=2.2.0
pymupdf>=1.24.0
python-dotenv>=1.0.0
deepeval>=1.0.0
aiosqlite>=0.20.0
asyncpg>=0.29.0
sqlalchemy[asyncio]>=2.0.0
psycopg2-binary>=2.9.9
google-cloud-alloydb-connector[pg8000]>=1.3.0
google-cloud-bigquery>=3.41.0
